consolidated data modules for flowfilerepostorage
